app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.2
cffi==1.17.1
click==8.2.0
colorama==0.4.6
deepdiff==8.5.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
motor==3.7.1
orderly-set==5.4.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.13.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
websockets==15.0.1
